Fix warning on Win64
git-svn-id: https://svn.osgeo.org/gdal/trunk@31791 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_vsil_win32.cpp
@@ -417,7 +417,7 @@ static void VSIWin32TryLongFilename(wchar_t*& pwszFilename)
         while( pwszFilename[nOffset + 0] == '.' && pwszFilename[nOffset + 1] == '.' &&
               (pwszFilename[nOffset + 2] == '/' || pwszFilename[nOffset + 2] == '\\') )
         {
-            size_t nCurDirLenBefore = nCurDirLen;
+            DWORD nCurDirLenBefore = nCurDirLen;
             while( nCurDirLen > 0 && pwszCurDir[nCurDirLen-1] != '\\' )
                 nCurDirLen --;
             if( nCurDirLen <= 2 )
